requirements.txt
requests
watchdog
pywin32pypdf
